requirements.txt
pygame
pyyaml
websocket-client